requirements.txt
fastapi
uvicorn[standard]
numpy
pandas
scikit-learn
imbalanced-learn
python-dotenv
